Feat: Added logic to pull timeblock objects in local storage and display in relevant time blocks.
Fix: Removed the alert when users saved an empty timeblock description. Makes sense, but in practical terms it blocks users from the ability to clear an timeblock description that needs to be cleared if cancelled etc. In future, we could return this alert and add a separate button to clear 1 or all descriptions.

Acceptance Criteria: When I refresh the page then the saved events persist.